refactor(controllers): reorganize routes and http app into controllers/http
