+++ controllers/http/v1/__init__.py
@@ -1,0 +1,3 @@
+from controllers.http.v1 import autoinfracao, consorcio, linha, recursos, veiculos
+
+__all__ = ["autoinfracao", "recursos", "veiculos", "linha", "consorcio"]
