+++ controllers/http/v1/schemas/veiculos.py
