--- routes/v1/__init__.py
@@ -1,3 +1,0 @@
-from routes.v1 import autoinfracao, consorcio, linha, recursos, veiculos
-
-__all__ = ["autoinfracao", "recursos", "veiculos", "linha", "consorcio"]
